Split the compress logic from upload
